remove otaclient_service module, merge it into the main module
